requirements/test.txt
# Test dependencies go here.

-r base.txt

# Pytest:
pytest==4.6.4  # pyup: <5.0.0
pytest-asyncio==0.10.0
pytest-cov==2.7.1
pytest-django==3.5.1
pytest-factoryboy==2.0.3
pytest-mock==1.10.4

coveralls==1.8.2
factory_boy==2.12.0
freezegun==0.3.12
tox==3.14.0

# Linting:
black==19.3b0
flake8==3.7.8
isort==4.3.21

# Docs:
Sphinx==2.2.0
sphinxcontrib-httpdomain==1.7.0
